+++ src/Repositories/GenericRepository.php
@@ -9,6 +9,8 @@ namespace dcardenasl\Ci4ApiCore\Repositories;
  *
  * A concrete implementation of BaseRepository for rapid transitioning
  * from direct Model usage to the Repository Pattern.
+ *
+ * @extends BaseRepository<object>
  */
 class GenericRepository extends BaseRepository
 {